Refactored datatable test suite
